v3
This creates a Python list from a set of specific ID's. These will become directories for getting .har type files, which can be parsed like a JSON file, and then inserted into MySQL. The .har (or JSON as it's treated) is a PageLoadTest, the generic test widely available for measuring speeds of URL domains.